requirements.txt
asgiref==3.10.0
Django==4.2.7
django-axes==6.0.0
django-ratelimit==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.5.0
django-filter>=23.0
